fix: ajustment receives req.cookies with refresh_token

+++ src/domain/users/application/usecases/refresh-token.usecase.ts
@@ -2,65 +2,77 @@ import { UseCase } from './usecase'
 import { TokenGateway } from '@/domain/users/application/gateways/token.gateway'
 import { TokenProvider } from '@/domain/ports/out/token'
 import { DateProvider } from '@/domain/ports/out/date'
-import { UserPresenters } from '@/domain/ports/in/user.present'
-import { RefreshTokenInputDto, RefreshTokenOutputDto } from '@/domain/dto/user/token-user.dto'
+import { RefreshTokenInputDto } from '@/domain/dto/user/token-user.dto'
 import { RefreshTokenNotExists } from '@/domain/errors/refesh-token-not-exists-error'
 import { env } from '@/config/env'
+import { AuthUserOutputDto } from '@/domain/dto/user/auth-user.dto'
+import { Hashing } from '@/domain/ports/out/hasher'
+import { InvalidCredentialsError } from '@/domain/errors/invalid-credentials-error'
 
 
 interface Payload {
   sub: string
 }
 
-export class RefreshTokenUseCase implements UseCase<RefreshTokenInputDto, RefreshTokenOutputDto> {
+export class RefreshTokenUseCase implements UseCase<RefreshTokenInputDto, AuthUserOutputDto> {
   private constructor(
     private readonly tokenGateway: TokenGateway,
-    private readonly userPresenters: UserPresenters,
     private readonly tokenProvider: TokenProvider,
+    private readonly encrypter: Hashing,
     private readonly dateProvider: DateProvider,
   ){}
 
   public static create(
     tokenGateway: TokenGateway,
-    userPresenters: UserPresenters,
     tokenProvider: TokenProvider,
+    encrypter: Hashing,
     dateProvider: DateProvider,
   ){
     return new RefreshTokenUseCase(
       tokenGateway,
-      userPresenters,
       tokenProvider,
+      encrypter,
       dateProvider,
     )
   }
 
-  async execute({ token }: RefreshTokenInputDto): Promise<RefreshTokenOutputDto> {
-    const { sub } = await this.tokenProvider.verifyRefreshToken(token) as Payload
+  async execute(refresh_token: RefreshTokenInputDto): Promise<AuthUserOutputDto> {
+    const { sub } = await this.tokenProvider.verifyRefreshToken(refresh_token.refresh_token) as Payload
     const user_id = sub
     
-    const userToken = await this.tokenGateway.findByUserIdAndRefreshToken(user_id, token)
+    const userToken = await this.tokenGateway.findByUserIdAndRefreshToken(user_id)
     if(!userToken){
       throw new RefreshTokenNotExists()
     }
 
+    const isRefreshTokenValid = await this.encrypter.compare(refresh_token.refresh_token, userToken.refresh_token)
+    if(!isRefreshTokenValid){
+      throw new InvalidCredentialsError()
+    }
+    
     await this.tokenGateway.deleteByTokenId(userToken.id.toString())
 
-    const newAccessToken = await this.tokenProvider.generateAccessToken({
+    const accessToken = await this.tokenProvider.generateAccessToken({
       sub: userToken.id.toString(),
     })
     
-    const refresh_token = await this.tokenProvider.generateRefreshToken({
+    const refreshToken = await this.tokenProvider.generateRefreshToken({
       sub: userToken.id.toString(),
     })
+
+    const refreshTokenHashed = await this.encrypter.hash(refreshToken)
     
     const expiresRefreshToken = this.dateProvider.addDays(env.EXPIRES_REFRESH_TOKEN_DAYS)
     
-    const newRefreshToken = await this.tokenGateway.create({
+    await this.tokenGateway.create({
       user_id: sub,
-      refresh_token,
+      refresh_token: refreshTokenHashed,
       expires_date: expiresRefreshToken,
     })
 
-    return await this.userPresenters.presentAuthUser(newAccessToken, newRefreshToken.refresh_token)
+    return {
+      token: accessToken,
+      refresh_token: refreshToken
+    }
   }
 }
